Perform multi-threading
Program spawns multiple threads and uses it to sum through all elements and find the maximum of elements in a matrix. Two techniques are employed in this project; one uses threads spawned at the beginning of the program and maintains all threads throughout (The worker thread program), and the other uses threads being spawned solely for the tasks they are created for.